+++ src/store/ui/actions.js
@@ -2,13 +2,7 @@ import Vue from "vue";
 import JSZip from "jszip";
 import { postMessage } from "../../utilities";
 
-import {
-  copyToClipboard,
-  exportFile,
-  Loading,
-  LocalStorage,
-  Dialog,
-} from "quasar";
+import { copyToClipboard, exportFile, LocalStorage, Dialog } from "quasar";
 import {
   prompt,
   notify,
@@ -19,16 +13,8 @@ import {
 } from "../../utilities";
 import { THEMES } from "../../themes";
 import { preload as preloadTakAnnotator } from "../../bots/tak-annotator";
-import { SHORTENER_SERVICE } from "../../constants";
 import { i18n } from "../../boot/i18n";
-import {
-  cloneDeep,
-  isArray,
-  isFunction,
-  isObject,
-  isString,
-  pick,
-} from "lodash";
+import { isArray, isFunction, isObject, isString, pick } from "lodash";
 import { TPStoPNG } from "tps-ninja";
 import {
   isThumbnailWorkerSupported,
@@ -474,72 +460,4 @@ export const GET_THUMBNAIL = ({ commit, state }, rawOptions) => {
       processThumbnailQueue();
     }
   });
-};
-
-export const GET_SHORT_URL = async ({ commit, state }, { game, options }) => {
-  if (!game) {
-    return "";
-  }
-  if (game.config.isOnline) {
-    return location.origin + "/game/" + game.config.id;
-  }
-  options = cloneDeep(options);
-
-  const ptn = game.ptn;
-  const params = {};
-
-  if ("name" in options) {
-    params.name = options.name || "";
-  } else if (game.name) {
-    params.name = game.name;
-  }
-
-  if (options.state) {
-    if (options.state === true) {
-      options.state = game.board;
-    }
-    if (options.state.targetBranch) {
-      params.targetBranch = options.state.targetBranch;
-    }
-    if (options.state.plyIndex >= 0) {
-      params.ply = String(options.state.plyIndex);
-      if (options.state.plyIsDone) {
-        params.ply += "!";
-      }
-    }
-  }
-
-  try {
-    const data = { ptn, params };
-    const hash = hashObject(data);
-    if (hash in state.shortLinks) {
-      return state.shortLinks[hash];
-    }
-    Loading.show();
-    const response = await fetch(SHORTENER_SERVICE, {
-      method: "POST",
-      headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json",
-      },
-      mode: "cors",
-      body: JSON.stringify(data),
-    });
-    if (!response.ok) {
-      const json = await response.json();
-      if (json && json.message) {
-        return notifyError(json.message);
-      } else {
-        return notifyError("HTTP-Error: " + response.status);
-      }
-    }
-    Loading.hide();
-    const url = await response.text();
-    commit("SET_SHORT_LINK", { hash, url });
-    return url;
-  } catch (error) {
-    Loading.hide();
-    notifyError(error);
-    return false;
-  }
 };